=str break out a few source implementations into separate files [backport]
(cherry picked from commit 084b1220ffafa543f5a7ad787faf64f21ed0e9e8)
